adjust model
