--- app/src/main/java/com/depromeet/housekeeper/model/ErrorResponse.kt
@@ -1,7 +1,0 @@
-package com.depromeet.housekeeper.model
-
-data class ErrorResponse(
-    val code: Int,
-    val errorMessage: String,
-    val referrerUrl: String
-)
