--- app/src/main/java/com/depromeet/housekeeper/model/request/BuildTeam.kt
@@ -1,5 +1,0 @@
-package com.depromeet.housekeeper.model.request
-
-data class BuildTeam(
-    val teamName: String
-)
