--- app/src/main/java/com/depromeet/housekeeper/model/request/ChoreList.kt
@@ -1,10 +1,0 @@
-package com.depromeet.housekeeper.model.request
-
-import kotlinx.serialization.Serializable
-
-@Serializable
-data class ChoreList(
-    val space: String,
-    val houseWorks: List<String>
-
-)
