--- app/src/main/java/com/depromeet/housekeeper/model/request/Chores.kt
@@ -1,8 +1,0 @@
-package com.depromeet.housekeeper.model.request
-
-import kotlinx.serialization.Serializable
-
-@Serializable
-data class Chores(
-    val houseWorks: List<Chore>,
-)
