--- app/src/main/java/com/depromeet/housekeeper/model/request/EditProfileModel.kt
@@ -1,7 +1,0 @@
-package com.depromeet.housekeeper.model.request
-
-data class EditProfileModel(
-    val memberName: String,
-    val profilePath: String,
-    val statusMessage: String
-)
