--- app/src/main/java/com/depromeet/housekeeper/model/request/JoinTeam.kt
@@ -1,5 +1,0 @@
-package com.depromeet.housekeeper.model.request
-
-data class JoinTeam(
-    val inviteCode: String
-)
