--- app/src/main/java/com/depromeet/housekeeper/model/request/Message.kt
@@ -1,7 +1,0 @@
-package com.depromeet.housekeeper.model.request
-
-data class Message(
-    val body: String,
-    val memberId: Int,
-    val title: String,
-)
