--- app/src/main/java/com/depromeet/housekeeper/model/request/Rule.kt
@@ -1,5 +1,0 @@
-package com.depromeet.housekeeper.model.request
-
-data class Rule(
-    val ruleName: String,
-)
