--- app/src/main/java/com/depromeet/housekeeper/model/request/SocialType.kt
@@ -1,5 +1,0 @@
-package com.depromeet.housekeeper.model.request
-
-data class SocialType(
-    val socialType: String
-)
