--- app/src/main/java/com/depromeet/housekeeper/model/request/Token.kt
@@ -1,5 +1,0 @@
-package com.depromeet.housekeeper.model.request
-
-data class Token(
-    val token: String
-)
