--- app/src/main/java/com/depromeet/housekeeper/model/request/UpdateChoreBody.kt
@@ -1,5 +1,0 @@
-package com.depromeet.housekeeper.model.request
-
-data class UpdateChoreBody(
-    val toBeStatus: Int
-)
