--- app/src/main/java/com/depromeet/housekeeper/model/request/UpdateMember.kt
@@ -1,6 +1,0 @@
-package com.depromeet.housekeeper.model.request
-
-data class UpdateMember(
-    val memberName: String,
-    val profilePath: String
-)
