--- app/src/main/java/com/depromeet/housekeeper/model/response/BuildTeamResponse.kt
@@ -1,11 +1,0 @@
-package com.depromeet.housekeeper.model.response
-
-import android.os.Parcelable
-import kotlinx.parcelize.Parcelize
-
-@Parcelize
-class BuildTeamResponse(
-    val inviteCode: String,
-    val teamId: Int
-) : Parcelable
-
