--- app/src/main/java/com/depromeet/housekeeper/model/response/CompleteHouseWork.kt
@@ -1,5 +1,0 @@
-package com.depromeet.housekeeper.model.response
-
-data class CompleteHouseWork(
-    val count: Int
-)
