--- app/src/main/java/com/depromeet/housekeeper/model/response/EditResponseBody.kt
@@ -1,6 +1,0 @@
-package com.depromeet.housekeeper.model.response
-
-data class EditResponseBody(
-    val code: Int,
-    val message: String
-)
