--- app/src/main/java/com/depromeet/housekeeper/model/response/GetInviteCode.kt
@@ -1,11 +1,0 @@
-package com.depromeet.housekeeper.model.response
-
-import android.os.Parcelable
-import kotlinx.parcelize.Parcelize
-
-@Parcelize
-data class GetInviteCode(
-    val inviteCode: String,
-    val inviteCodeExpirationDateTime: String,
-    val teamName: String
-) : Parcelable
