--- app/src/main/java/com/depromeet/housekeeper/model/response/HouseWorkCreateResponse.kt
@@ -1,10 +1,0 @@
-package com.depromeet.housekeeper.model.response
-
-import android.os.Parcelable
-import com.depromeet.housekeeper.model.HouseWork
-import kotlinx.parcelize.Parcelize
-
-@Parcelize
-data class HouseWorkCreateResponse(
-    val houseWorks: List<HouseWork>
-) : Parcelable
