--- app/src/main/java/com/depromeet/housekeeper/model/response/LoginResponse.kt
@@ -1,17 +1,0 @@
-package com.depromeet.housekeeper.model.response
-
-import android.os.Parcelable
-import kotlinx.parcelize.Parcelize
-
-
-@Parcelize
-data class LoginResponse(
-    val accessToken: String,
-    val accessTokenExpireTime: String,
-    val hasTeam: Boolean,
-    val isNewMember: Boolean,
-    val memberName: String?,
-    val memberId: Int,
-    val refreshToken: String,
-    val refreshTokenExpireTime: String
-) : Parcelable
