--- app/src/main/java/com/depromeet/housekeeper/model/response/ProfileData.kt
@@ -1,7 +1,0 @@
-package com.depromeet.housekeeper.model.response
-
-data class ProfileData(
-    val memberName: String,
-    val profilePath: String,
-    val statusMessage: String = "",
-)
