--- app/src/main/java/com/depromeet/housekeeper/model/response/ProfileImages.kt
@@ -1,5 +1,0 @@
-package com.depromeet.housekeeper.model.response
-
-data class ProfileImages(
-    val bigImageList: List<String>,
-)
