--- app/src/main/java/com/depromeet/housekeeper/model/response/Response.kt
@@ -1,6 +1,0 @@
-package com.depromeet.housekeeper.model.response
-
-data class Response(
-    val code: Int,
-    val message: String
-)
