--- app/src/main/java/com/depromeet/housekeeper/model/response/RuleResponse.kt
@@ -1,6 +1,0 @@
-package com.depromeet.housekeeper.model.response
-
-data class RuleResponse(
-    val ruleId: Int,
-    val ruleName: String
-)
