--- app/src/main/java/com/depromeet/housekeeper/model/response/RuleResponses.kt
@@ -1,6 +1,0 @@
-package com.depromeet.housekeeper.model.response
-
-data class RuleResponses(
-    val ruleResponseDtos: List<RuleResponse>,
-    val teamId: Int
-)
