--- app/src/main/java/com/depromeet/housekeeper/model/response/TeamUpdateResponse.kt
@@ -1,6 +1,0 @@
-package com.depromeet.housekeeper.model.response
-
-data class TeamUpdateResponse(
-    val teamId: Int,
-    val teamName: String
-)
