--- app/src/main/java/com/depromeet/housekeeper/model/response/UpdateChoreResponse.kt
@@ -1,6 +1,0 @@
-package com.depromeet.housekeeper.model.response
-
-data class UpdateChoreResponse(
-    val houseWorkId: Int,
-    val success: Boolean
-)
